+++ src/ui/guitools/chartwrapper.cpp
@@ -228,6 +228,8 @@ QList<QPointer<QtCharts::QScatterSeries>> ChartWrapper::CloneSeries() const
 {
     QList<QPointer<QtCharts::QScatterSeries>> series;
     for (int i = 0; i < m_stored_series.size(); ++i) {
+        if (!m_stored_series[i]->isVisible())
+            continue;
         QtCharts::QScatterSeries* serie = new QtCharts::QScatterSeries();
         serie->append(m_stored_series[i]->points());
         serie->setColor(m_stored_series[i]->color());